10.12

+++ core/src/com/unimelb/swen30006/partc/core/World.java
@@ -86,7 +86,10 @@ public class World implements ISteppable {
 		// Controllers and cars
 		this.controllers = new Controller[1];
 		this.cars = new Car[1];
-		this.cars[0] = new Car(new Point2D.Double(80,140), 6, 10, Color.CORAL, 25f, 50f, 6f );
+		//this.cars[0] = new Car(new Point2D.Double(80,140), 6, 10, Color.CORAL, 25f, 50f, 6f );
+		
+		this.cars[0] = new Car(new Point2D.Double(160,160), 6, 10, Color.CORAL, 25f, 50f, 6f );
+		
 		//this.controllers[0] = new KeyboardController(cars[0]);
 		this.controllers[0] = new AIController(cars[0], new Planning(cars[0], roads), new PerceptionForTesting(this));
